oracle & postgresql dialect
