+++ tests/dialects/test_redshift_dialect.py
@@ -64,6 +64,8 @@ def test_base_encode():
                 "sortkey": {"keys": [], "type": None},
                 "table_name": "sales",
                 "tablespace": None,
+                "row_format": None,
+                "stored_as": None,
                 "temp": False,
             }
         ],
@@ -252,6 +254,8 @@ def test_distkey_sortkey():
                 "primary_key": ["salesid"],
                 "schema": None,
                 "encode": None,
+                "row_format": None,
+                "stored_as": None,
                 "sortkey": {"keys": ["listid", "sellerid"], "type": "compound"},
                 "table_name": "sales",
                 "tablespace": None,
@@ -296,6 +300,8 @@ def test_distyle():
                 "index": [],
                 "encode": None,
                 "partitioned_by": [],
+                "row_format": None,
+                "stored_as": None,
                 "primary_key": [],
                 "schema": None,
                 "sortkey": {"keys": [], "type": None},
@@ -355,6 +361,8 @@ def test_encode_for_full_table():
                 "schema": None,
                 "sortkey": {"keys": [], "type": None},
                 "table_name": "t2",
+                "row_format": None,
+                "stored_as": None,
                 "tablespace": None,
                 "temp": False,
             }
@@ -481,9 +489,13 @@ def test_interleaved_sortkey_also_ok():
                 "distkey": None,
                 "diststyle": "all",
                 "encode": None,
+                "row_format": None,
+                "stored_as": None,
                 "index": [],
                 "partitioned_by": [],
                 "primary_key": [],
+                "row_format": None,
+                "stored_as": None,
                 "schema": None,
                 "sortkey": {
                     "keys": ["c_custkey", "c_city", "c_mktsegment"],
@@ -562,6 +574,8 @@ def test_create_temp_table():
                 "sortkey": {"keys": [], "type": None},
                 "table_name": "tempevent",
                 "tablespace": None,
+                "row_format": None,
+                "stored_as": None,
                 "temp": True,
             }
         ],
@@ -604,6 +618,8 @@ def test_like_in_parath():
                 "table_name": "tempevent",
                 "tablespace": None,
                 "temp": True,
+                "row_format": None,
+                "stored_as": None,
             }
         ],
         "types": [],
