+++ tests/dialects/test_spark_sql.py
@@ -52,6 +52,8 @@ def test_spark_sql_using():
                 ],
                 "index": [],
                 "partitioned_by": [],
+                "row_format": None,
+                "stored_as": None,
                 "primary_key": [],
                 "schema": None,
                 "table_name": "student",
@@ -170,7 +172,9 @@ def test_spark_sql_partitioned_by_function():
                 "index": [],
                 "partitioned_by": [{"args": "(b)", "func_name": "months"}],
                 "primary_key": [],
+                "row_format": None,
                 "schema": None,
+                "stored_as": None,
                 "table_name": "a",
                 "tablespace": None,
                 "location": "'s3://tables/a'",
